Add device state infrastructure
In order to migrate between virtiofsd instances, we need to transfer our
internal state from the source instance to the destination instance
(basically all content of the PassthroughFs object).

To achieve this, vhost-user provides a device state feature, which has
the front end request internal state from the source back end during
migration, receive it, transfer it to the destination, and provide it to
the destination back end instance there.  This way, the back end (i.e.
virtiofsd) does not need to find and connect to its destination
instance, and all required state can be contained within a single
migration stream maintained by the front end.

This feature comes with two messages: SET_DEVICE_STATE_FD, and
CHECK_DEVICE_STATE.  The former establishes a communication channel over
which to transfer the state data, the latter checks for success after
the transfer is complete.  We implement these in VhostUserFsBackend on
the basis of a more expressive trait, which is the
SerializableFileSystem:
- When migration is initiated (which is signalled by the front end
  setting the LOG_ALL feature bit), we begin to prepare for migration.
  This may involve gathering some data, so it can take a while.
  SerializableFileSystem::prepare_serialization() is launched in a new
  thread, so it can takes its time to do this.
  This function takes an AtomicBool that is supposed to be checked
  regularly by the prepare_serialization() implementation, and if set,
  cancel the preparation and exit the thread.  We set that AtomicBool
  when the front end stops migration without asking for our device state
  (i.e. migration is cancelled).
- When the front end does ask for our device state, we serialize the
  prepared data via SerializableFileSystem::serialize(), and send it to
  the front end.  This too is run in a thread, because the
  SET_DEVICE_STATE_FD call must not block.
- When the front end supplies device state (on the destination side), we
  deserialize it via SerializableFileSystem::deserialize_and_apply().
  This too is run a thread.

SerializableFileSystem will be implemented in a later commit.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>