Add support for dirty memory logging
Add support for SET_LOG_BASE handling: This allows the front end to
establish a bitmap in which we have to log all memory pages that we
modify.  Migration requires such a bitmap to determine which memory
pages need to be (re-)transferred to the destination, and because the VM
can only (reliably) track memory accesses done by the guest, a vhost
back end needs to track its own accesses separately.

Luckily, the vhost crates already provide do all the work for us, all we
need to do is ask them to do this work by:
1. Requiring an actual bitmap to be attached to our back end
   implementation (i.e. VhostUserFsBackend) by changing the Bitmap type
   parameter from () to BitmapMmapRegion (which is an mmap-based bitmap)
2. Telling the front end we support the LOG_SHMFD and LOG_ALL features

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>