Hide InodeStore locking
Instead of having users of InodeStore lock and unlock it as needed,
rename it to InodeStoreInner, hide it in inode_store.rs, and create a
new InodeStore struct that contains an `Arc<RwLock<InodeStoreInner>>`.

Besides making using the InodeStore a bit less cumbersome this way,
this will be kind of necessary when introducing a type for strong inode
references (i.e. ones that count against the inode store refcount):
Dropping such a strong reference should automatically release the inode
when the refcount drops to 0, which requires locking the InodeStore.
Such references must therefore not be dropped while the store is locked,
or a deadlock will occur.  To see that this will not happen, we should
concentrate all InodeStore locking in a single place, specifically
inode_store.rs.

Use this opportunity to move some functions from the PassthroughFs
implementation into inode_store.rs: forget_one(), claim_inode(), and the
code for adding a new inode from do_lookup() into a new function
get_or_insert().  As for the latter, also drop the comment about a
"race" regarding the use of PassthroughFs.next_inode: First, this
comment evoked a sense of hazard, when really, everything is completely
fine.  Second, it called eventually unused IDs "wasted", when really,
they are just numbers that are skipped -- nobody has ever required or
intended these IDs to be consecutive, the only thing that is important
is that the ID is unique, and that it obviously is by virtue of using an
atomic operation (unless we overflow, but that's a different story).
Basically, the comment was helpful before 38ecb626, when there was a
real problem here, but that commit thankfully fixed that problem, and
now the comment is more confusing than helpful, so remove it.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>