Put open_root_node() into own function
Move the code from PassthroughFs::init() that opens the root node into a
dedicated function.  We are going to want to use this function when
deserializing incoming migration data to emulate this init code (the
guest is not going to send an INIT after migration, so the
deserialization code will have to initialize the filesystem based on the
incoming device state).

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>