Add ReadDir::new_no_seek()
Seeking in directories is reportedly apparently broken on extfs, so
provide a new version of ReadDir that does not seek before reading
entries.  For this to work, the caller must ensure that they do not use
the associated file descriptor for anything but ReadDir, so this
function is marked as unsafe.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>