Define our serialized device state
This patch adds the structures that we are going to use to represent our
internal device state (the PassthroughFs object) on the wire, and a
dependency on serde to make them serializable.

The recursive dependencies serde brings (serde, serde_derive,
proc-macro2, quote, syn) are, according to cargo-license, all licensed
under Apache-2.0 OR MIT, which is the same as most of our existing
dependencies, so that should be fine.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>