Implement preserialization/premigration
Add functionality and structures to prepare for serialization.

Each inode in our store must be accompanied by data that will allow the
destination to find it, and each handle must be accompanied by data that
will allow the destination to open it.

The latter is trivial to obtain (just the open(2) flags), so can be kept
all the time.  The former may be more complicated: In its current form,
inodes will be represented by a reference to their parent directory
inode and their filename.  We do not want to keep these around all the
time because external parties may rename files and thus invalidate this
information without us knowing.  Therefore, when migrating, the location
information we have for inodes should be as fresh as possible, which is
why we start to collect it only once migration begins, i.e. in the
prepare_serialization() function.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>