Implement device state deserialization
Deserialization is a bit trickier than serialization, because we not
only need to turn the binary data back into its structured
representation (this is trivial, because serde + postcard do all the
work), but also need to apply the represented state to an existing
PassthroughFs object.  That means we need to locate all inodes and
populate the inode store, and open all handles to populate the handle
map.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>