Add get_path_by_fd()/printable_fd() functions
get_path_by_fd() function reads the corresponding symlink in
/proc/self/fd to find out what path the kernel thinks the corresponding
inode has.

Through printable_fd(), this can be used for error and info messages
where we only have an FD and want to give the user a better idea of what
that FD probably is.

On the other hand, for migration, we can also use get_path_by_fd() in
place of or to complement our current method (iterating through the
shared directory) of finding corresponding paths for FDs.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>