Allow explicitly invalid inodes and handles
During migration, it is possible that the destination side fails to open
some inodes.  One way to react to this is to abort migration and force
qemu to continue on the source side.  The other is to remember these
inodes as invalid and return I/O errors to the guest whenever it tries
to access them.

This commit introduces such invalid inodes handles: For inodes, there is
a new `Invalid` variant for `FileOrHandle`, and for handles, the file in
`HandleData` can be invalid instead of a `File`.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>