Introduce --migration-confirm-paths
This option has the source double-check every inode it serializes (via
statx(), or the file handle when using --migration-verify-handles)
whether it is still the inode we mean.  If not, we try to correct our
info via data from /proc/self/fd.

This allows us to notice changes brought about by other programs
accessing the shared directory, but it requires doing I/O during the
migration switch-over phase, i.e. when both source and destination are
not running, which is not ideal.

Note that this commit brings an incompatible change to the v1 migration
format, so must be merged in the same release as the patches introducing
that format.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>