Introduce --migration-on-error switch
This new switch will allow users to define what happens on the migration
destination when it finds itself unable to open some inode: Abort the
whole migration process (having qemu return control to the source side),
or continue, but return errors to the guest whenever it tries to access
any of the inodes that were failed to be opened.

Right now, this switch of course does not do anything, because migration
is not fully implemented yet.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>