Introduce --migration-verify-handles
This option has the source side generate a file handle for every inode
it sends to the destination.  The destination is not supposed to use
this file handle to open/locate the inode (it's still path-based, hence
"find-paths"), but instead to verify that the path it has opened refers
to the same inode by generating its file handle and verifying that both
are equal.

Of course, this only works if source and destination use the same shared
directory on the same filesystem, but this can also be a network
filesystem with source and destination operating on different hosts.

It is possible that some filesystems generate different handles for the
same file under various circumstances, but that is just something that
needs to be tested.

In any case, double-checking in the destination is important for a
multi-writer migration case, i.e. where external programs have write
access to the shared directory while virtiofsd is migrating.  In this
case, these external programs can rename and replace inodes after the
source has last checked the association between path and inode, and
before the destination has opened the path.  Verifying the file handle
in the destination will make it detect such a case.

Note that requiring this mode for safe multi-writer migration, and thus
requiring source and destination to share the same directory on the same
filesystem is not an actual limitation: Basically, any reasonable form
of multi-writer migration (right now) is only imaginable under this
constraint anyway.

Right now, this switch of course does not do anything, because migration
is not fully implemented yet.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>