Split try_lookup() off of do_lookup()
PassthroughFs::do_lookup() looks up the given filename under the given
parent inode (directory), and if it is not yet in our inode store, adds
it to it.

We are going to want another version that does not do the second part,
i.e. one that only looks up a filename in a directory, but does not add
it to our inode store if it does not exist.  Split off the first part of
do_lookup() into try_lookup_implementation(), and use that as the basis
of both try_lookup() and do_lookup().

Note that do_lookup() retains it interface, i.e. takes an `Inode`
parameter, whereas try_lookup() takes an `&InodeData`.  For the former,
we want to keep the current interface for minimal changes; for the
latter, the caller we are going to introduce later will have the
respective `&InodeData` anyway, which makes things simpler.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>