Add type for strong inode references
InodeData.refcount is currently only used to track the guest's
references, i.e. when to drop the inode when forget_one() is called.
But for path-based migration, we will need a system for tracking strong
references, too, so introduce a new type StrongInodeReference that is a
reference to an inode in our store, and "owns" an increment to its
refcount.

For migration, we need to tell the destination instance for every single
inode how it can find that inode in its shared directory.  When using
path-based migration, inodes are defined by a reference to their parent
(directory) node, and their filename.  These parent references
eventually reach a root, specifically the shared directory root node.

To ensure that all those parent references are valid and part of our
serialized state, they must be strong references, so they will not
disappear until we are done with migration.  That is what this new type
will be used for.

InodeStore::claim_inode() (PassthroughFs::claim_inode() before HEAD^)
has basically already established this concept of strong references, so
have it return a StrongInodeReference.  The same applies to
InodeStore::get_or_insert().

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>