Shorten `std::result::Result` where possible
Unless overloaded, `Result` always refers to `std::result::Result`, so
we do not have to specify the full form (unless in places where we do).

It should still be simple to distinguish between the standard `Result`
and module-specific `Result` types, given that the former has two
generic type parameters, and the latter ones only have one.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>