Hide clippy warning in readdir
There is a new clippy warning in our readdir handler: Clippy complains
that we use a write lock to gain access to an object that we then do not
mutate, i.e. we should be using a read lock here.  However, the existing
comment seems pretty clear that we do in fact mutate some of the
object's state, just outside of our process's scope, and thus invisible
to Rust and clippy.

Keep the write lock, just make clippy ignore it.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>