Add '--shared-dir' as required if '-o' is missing
If virtiofsd is run without any option an error is shown
with required arguments. Since, we accept both '--shared-dir'
and '-o source=' we cannot define '--shared-dir' as required.

Let's set it as required unless the '-o' option is provided.
This is not perfect, since it will not show '--shared-dir'
as required if any  '-o' is provided, not only '-o source',
but I think is enough because '-o' is deprecated, and it will
be removed eventually.

Fixes: https://issues.redhat.com/browse/RHEL-7348
Signed-off-by: German Maglione <gmaglione@redhat.com>