Add ErrorContext, ResultErrorContext traits
The new `ErrorContext` and `ResultErrorContext` traits provide easy ways
to give context to a potential error raised by some function,
specifically by providing a string that is prepended.

In Rust, one is often inclined to handle errors via a simple `?`, but
any error message that is generated multiple layers deep is sure to lack
some context that would be important to users, so passing such messages
untouched through the stack is going to result in subpar and confusing
error messages.

The `ResultErrorContext` trait provides a relatively non-obtrusive way
to add such context before passing on errors.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>