Force-call DESTROY on INIT
It is possible we receive two INITs without a DESTROY in between; e.g.
by virtue of a not very well behaving guest, or when the VM is hard
reset.  We must not allow old FUSE state to persist through INIT, but
currently, we only clear it when receiving an explicit DESTROY.

So clear our state on INIT, too.

This fixes an issue introduced in 8e1c21d0121 (“Hide InodeStore
locking”): That commit changed the way we add the root node to our inode
store, specifically instead of force-overwriting any potential previous
instance, INIT will now return an error if such a previously existing
root node would be overwritten.  By clearing all of our state before
executing INIT, we rule that out.

Tested-by: Nathan Chancellor <nathan@kernel.org>
Fixes: #163
Fixes: #166
Fixes: 8e1c21d0121 (“Hide InodeStore locking”)
Signed-off-by: Hanna Czenczek <hreitz@redhat.com>