Fix clippy doc_lazy_continuation warnings
Clippy has learned to complain about subsequent lines of list items in
comments not being indented correctly.  What an interesting thing for it
to care about, but it actually is kind of nice; fix those places.

(For add_cap_to_eff(), use that chance to fix up a singular to be a
plural form, unify the line lengths in that comment, and fix up double
spacing.)

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>