Make preserialization infallible
We do not want preserialization to fail, it is supposed to be a
best-effort process that may yield an arbitrary number of non-migratable
inodes.  The destination is supposed to decide on how to deal with it,
not the source side (where preserialization runs).

Make it infallible, and handle all potential errors internally (by
printing warnings, mostly).

This fixes a bug where preserialization aborts if the shared directory
contains entries that are not accessible to virtiofsd.  Previously, that
would cause a hard error, aborting preserialization and thus the whole
serialization process; when actually, because those entries are
inaccessible, we cannot have them in our inode store, so nothing needs
to be migrated there anyway.

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>