Fix snapshotting without F_LOG_ALL
When using savevm or migrating to file with the background-snapshot flag
set, qemu will not set the F_LOG_ALL flag.  We use that flag as the
signal to start preserialization, so without it ever being set, we won’t
do preserialization, which dooms serialization to not be able to
serialize anything, really.

If we get to the serialization step (vhost-user’s SET_DEVICE_STATR_FD
command) and do not find a (previously) active premigration thread,
print a warning, and do preserialization then.

Reported-by: Lili Zhu <lizhu@redhat.com>
Fixes: https://issues.redhat.com/browse/RHEL-50542
Signed-off-by: Hanna Czenczek <hreitz@redhat.com>