Add prefix to invalid inode/handle errors
When generating errors for accessing invalid inodes or handles, we
should give context.

The errors contained in those invalid inodes or handles often refer to
“the migration source”, which may be confusing to see during normal VM
operation (i.e. while not migrating); and even more confusing when they
appear during the next migration, e.g.:

[ERROR virtiofsd::passthrough::device_state::preserialization::proc_paths]
Inode X: Failed to get path from /proc/self/fd: Inode X is not the same
inode as in the migration source: File handle differs: X != Y

In this case (with --migration-mode=proc-paths), the proc_paths module
cannot generate a path for this inode because it is inaccessible because
--migration-verify-handles invalidated it during the previous migration.
Speaking of “the migration source” is really confusing, though.

With this patch, the error message becomes instead:

[ERROR virtiofsd::passthrough::device_state::preserialization::proc_paths]
Inode X: Failed to get path from /proc/self/fd: Inode is invalid because
of an error during the preceding migration, which was: Inode X is not
the same inode as in the migration source: File handle differs: X != Y

Signed-off-by: Hanna Czenczek <hreitz@redhat.com>