Clear S_ISGID explicitly in setxattr handler
Guest kernel in accordance with a FUSE protocol may
ask us to clear S_ISGID for the inode. To achieve that
we are trying to simulate a setup when host kernel will
do that for us implicitly (we drop CAP_FSETID and change
GID to achieve this). But this is a bit fragile and can
break in some circumstances. Let's just use stat()+chmod()
and clean S_ISGID bit explicitly. In this case, we don't
need all this capabilities+setuid tricks at all.

Tested with xfstest generic/375 test case.

See also:
https://gitlab.com/qemu-project/qemu/-/commit/f1aa1774dfb79ce0546d9fbc1ce05cb00550c600

Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>